Change floats for doubles for higher precision

+++ src/FusionEKF.cpp
@@ -64,11 +64,11 @@ void FusionEKF::ProcessMeasurement(const MeasurementPackage &measurement_pack)
     if (measurement_pack.sensor_type_ == MeasurementPackage::RADAR)
     {
       // Read out the Radara measurement components
-      float rho = measurement_pack.raw_measurements_[0];
-      float phi = measurement_pack.raw_measurements_[1];
+      double rho = measurement_pack.raw_measurements_[0];
+      double phi = measurement_pack.raw_measurements_[1];
 
-      float px = rho * cos(phi);
-      float py = rho * sin(phi);
+      double px = rho * cos(phi);
+      double py = rho * sin(phi);
 
       // Do not use rho dot in init phase since
       // radar measurement does not contain enough information to determin
@@ -117,17 +117,17 @@ void FusionEKF::ProcessMeasurement(const MeasurementPackage &measurement_pack)
   }
 
   // Prediction
-  float dt = (measurement_pack.timestamp_ - previous_timestamp_) / 1000000.0;
+  double dt = (measurement_pack.timestamp_ - previous_timestamp_) / 1000000.0;
   previous_timestamp_ = measurement_pack.timestamp_;
 
   ekf_.F_(0, 2) = dt;
   ekf_.F_(1, 3) = dt;
 
-  float dt_2 = dt * dt;
-  float dt_3 = dt_2 * dt;
-  float dt_4 = dt_3 * dt;
-  float dt_4_div_4 = dt_4 / 4.0;
-  float dt_3_div_2 = dt_3 / 2.0;
+  double dt_2 = dt * dt;
+  double dt_3 = dt_2 * dt;
+  double dt_4 = dt_3 * dt;
+  double dt_4_div_4 = dt_4 / 4.0;
+  double dt_3_div_2 = dt_3 / 2.0;
 
   ekf_.Q_ = MatrixXd(4, 4);
   ekf_.Q_ << dt_4_div_4 * noise_ax_, 0.0, dt_3_div_2 * noise_ax_, 0.0,
